Create PasswordGenerator.php
The script defines a function called generatePassword that takes a length parameter. This function generates a random password of the specified length using a set of characters.

The $characters variable stores a string of characters that can be used to generate the password. It includes lowercase and uppercase letters, numbers, and special characters.

The generatePassword function initializes an empty string called $password and calculates the total number of characters in the $characters string.

The function then enters a loop that runs $length number of times. In each iteration, a random index is generated using the mt_rand function, which is a random number generator. The index is used to select a character from the $characters string, which is appended to the $password string.

After the loop completes, the generated password is returned.

The script then checks if the form has been submitted by checking the $_POST['generate'] variable. If the form has been submitted, it retrieves the desired password length from the $_POST['length'] variable. If no length is specified, the default length is set to 8.

The generatePassword function is called with the specified length, and the resulting password is stored in the $password variable.

The HTML code follows, which includes a form with an input field for the password length and a submit button. The initial value of the length input field is set to the value submitted in case the form is reloaded.

If a password has been generated (i.e., $password is not empty), it is displayed below the form.